apps/api/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
supabase==2.3.5
pydantic==1.10.18
httpx>=0.25.0,<0.26
python-jose[cryptography]==3.3.0
pypdf==4.0.2
python-multipart==0.0.9
reportlab==4.0.8
